refactor(types): extract cosmos-api
