+++ packages/orchestration/src/types.ts
@@ -1,26 +1,23 @@
 // Ambients
 import '@agoric/zoe/exported.js';
 
-import type { AnyJson } from '@agoric/cosmic-proto';
-import type {
-  Delegation,
-  Redelegation,
-  UnbondingDelegation,
-} from '@agoric/cosmic-proto/cosmos/staking/v1beta1/staking.js';
-import type { TxBody } from '@agoric/cosmic-proto/cosmos/tx/v1beta1/tx.js';
 import { MsgTransferResponse } from '@agoric/cosmic-proto/ibc/applications/transfer/v1/tx.js';
-import type { Amount, Brand, Payment, Purse } from '@agoric/ertp/exported.js';
-import type { Port } from '@agoric/network';
+import type { Amount, Brand, Payment } from '@agoric/ertp/exported.js';
 import type { Timestamp } from '@agoric/time';
 import type { IBCConnectionID } from '@agoric/vats';
 import type { LocalChainAccount } from '@agoric/vats/src/localchain.js';
-import type {
-  LocalIbcAddress,
-  RemoteIbcAddress,
-} from '@agoric/vats/tools/ibc-utils.js';
 import type { EthChainInfo } from './ethereum-api.js';
 import type { ICQConnection } from './exos/icqConnectionKit.js';
+import type {
+  CosmosChainInfo,
+  Proto3JSONMsg,
+  StakingAccountQueries,
+  StakingAccountActions,
+  IcaAccount,
+  LiquidStakingMethods,
+} from './cosmos-api.js';
 
+export type * from './cosmos-api.js';
 export type * from './ethereum-api.js';
 export type * from './exos/chainAccountKit.js';
 export type * from './exos/icqConnectionKit.js';
@@ -34,9 +31,7 @@ export type * from './vat-orchestration.js';
 export type KnownChains = {
   stride: {
     info: CosmosChainInfo;
-    methods: {
-      liquidStake: (amount: AmountArg) => Promise<void>;
-    };
+    methods: LiquidStakingMethods;
   };
   cosmos: { info: CosmosChainInfo; methods: {} };
   agoric: {
@@ -55,9 +50,6 @@ export type KnownChains = {
   celestia: { info: CosmosChainInfo; methods: {} };
   osmosis: { info: CosmosChainInfo; methods: {} };
 };
-
-/** A helper type for type extensions. */
-export type TypeUrl = string;
 
 /**
  * A denom that designates a token type on some blockchain.
@@ -78,10 +70,6 @@ export type Denom = string; // ibc/... or uist
  */
 export type DenomArg = Brand | Denom;
 
-export type Proto3JSONMsg = {
-  '@type': TypeUrl;
-  value: Record<string, unknown>;
-};
 /** An address on some blockchain, e.g., cosmos, eth, etc. */
 export type ChainAddress = {
   /** e.g. 1 for Ethereum, agoric-3 for Agoric, cosmoshub-4 for Cosmos */
@@ -89,13 +77,6 @@ export type ChainAddress = {
   address: string;
   // TODO what's the right way to scope the address? it's not chainId
   addressEncoding: 'bech32' | 'ethereum';
-};
-
-/** An address for a validator on some blockchain, e.g., cosmos, eth, etc. */
-export type CosmosValidatorAddress = ChainAddress & {
-  // TODO document why this is the format
-  address: `${string}valoper${string}`;
-  addressEncoding: 'bech32';
 };
 
 /** Details for setup will be determined in the implementation. */
@@ -164,36 +145,6 @@ export type OrchestrationHandlerMaker = <C extends object>(
   fn: (orc: Orchestrator, ctx2: C, ...args) => object,
 ) => (...args) => object;
 
-/**
- * Info for a Cosmos-based chain.
- */
-export type CosmosChainInfo = {
-  chainId: string;
-  ibcConnectionInfo: {
-    id: string; // e.g. connection-0
-    client_id: string; // '07-tendermint-0'
-    state: 'OPEN' | 'TRYOPEN' | 'INIT' | 'CLOSED';
-    counterparty: {
-      client_id: string;
-      connection_id: string;
-      prefix: {
-        key_prefix: string;
-      };
-    };
-    versions: { identifier: string; features: string[] }[];
-    delay_period: bigint;
-  };
-  icaEnabled: boolean;
-  icqEnabled: boolean;
-  pfmEnabled: boolean;
-  ibcHooksEnabled: boolean;
-  /**
-   *
-   */
-  allowedMessages: TypeUrl[];
-  allowedQueries: TypeUrl[];
-};
-
 export type ChainInfo = CosmosChainInfo | EthChainInfo;
 
 // marker interface
@@ -234,147 +185,13 @@ export interface Chain<C extends keyof KnownChains> {
 }
 
 /**
- * Low level object that supports queries and operations for an account on a remote chain.
- */
-export interface ChainAccount {
-  /**
-   * @returns the address of the account on the remote chain
-   */
-  getAddress: () => ChainAddress;
-  /**
-   * Submit a transaction on behalf of the remote account for execution on the remote chain.
-   * @param msgs - records for the transaction
-   * @returns acknowledgement string
-   */
-  executeTx: (msgs: Proto3JSONMsg[]) => Promise<string>;
-  /**
-   * Submit a transaction on behalf of the remote account for execution on the remote chain.
-   * @param {AnyJson[]} msgs - records for the transaction
-   * @param {Partial<Omit<TxBody, 'messages'>>} [opts] - optional parameters for the Tx, like `timeoutHeight` and `memo`
-   * @returns acknowledgement string
-   */
-  executeEncodedTx: (
-    msgs: AnyJson[],
-    opts?: Partial<Omit<TxBody, 'messages'>>,
-  ) => Promise<string>;
-  /** deposit payment from zoe to the account*/
-  deposit: (payment: Payment) => Promise<void>;
-  /** get Purse for a brand to .withdraw() a Payment from the account */
-  getPurse: (brand: Brand) => Promise<Purse>;
-  /**
-   * Close the remote account
-   */
-  close: () => Promise<void>;
-  /* transfer account to new holder */
-  prepareTransfer: () => Promise<Invitation>;
-  /** @returns the address of the remote channel */
-  getRemoteAddress: () => RemoteIbcAddress;
-  /** @returns the address of the local channel */
-  getLocalAddress: () => LocalIbcAddress;
-  /** @returns the port the ICA channel is bound to */
-  getPort: () => Port;
-}
-
-export interface StakingAccountQueries {
-  /**
-   * @returns all active delegations from the account to any validator (or [] if none)
-   */
-  getDelegations: () => Promise<Delegation[]>;
-
-  /**
-   * @returns the active delegation from the account to a specific validator. Return an
-   * empty Delegation if there is no delegation.
-   */
-  getDelegation: (validator: CosmosValidatorAddress) => Promise<Delegation>;
-
-  /**
-   * @returns the unbonding delegations from the account to any validator (or [] if none)
-   */
-  getUnbondingDelegations: () => Promise<UnbondingDelegation[]>;
-
-  /**
-   * @returns the unbonding delegations from the account to a specific validator (or [] if none)
-   */
-  getUnbondingDelegation: (
-    validator: CosmosValidatorAddress,
-  ) => Promise<UnbondingDelegation>;
-
-  getRedelegations: () => Promise<Redelegation[]>;
-
-  getRedelegation: (
-    srcValidator: CosmosValidatorAddress,
-    dstValidator?: CosmosValidatorAddress,
-  ) => Promise<Redelegation>;
-
-  /**
-   * Get the pending rewards for the account.
-   * @returns the amounts of the account's rewards pending from all validators
-   */
-  getRewards: () => Promise<ChainAmount[]>;
-
-  /**
-   * Get the rewards pending with a specific validator.
-   * @param validator - the validator address to query for
-   * @returns the amount of the account's rewards pending from a specific validator
-   */
-  getReward: (validator: CosmosValidatorAddress) => Promise<ChainAmount[]>;
-}
-export interface StakingAccountActions {
-  /**
-   * Delegate an amount to a validator. The promise settles when the delegation is complete.
-   * @param validator - the validator to delegate to
-   * @param amount  - the amount to delegate
-   * @returns void
-   */
-  delegate: (
-    validator: CosmosValidatorAddress,
-    amount: AmountArg,
-  ) => Promise<void>;
-
-  /**
-   * Redelegate from one delegator to another.
-   * Settles when the redelegation is established, not 21 days later.
-   * @param srcValidator - the current validator for the delegation.
-   * @param dstValidator - the validator that will receive the delegation.
-   * @param amount - how much to redelegate.
-   * @returns
-   */
-  redelegate: (
-    srcValidator: CosmosValidatorAddress,
-    dstValidator: CosmosValidatorAddress,
-    amount: AmountArg,
-  ) => Promise<void>;
-
-  /**
-   * Undelegate multiple delegations (concurrently). To delegate independently, pass an array with one item.
-   * Resolves when the undelegation is complete and the tokens are no longer bonded. Note it may take weeks.
-   * The unbonding time is padded by 10 minutes to account for clock skew.
-   * @param {Delegation[]} delegations - the delegation to undelegate
-   */
-  undelegate: (delegations: Delegation[]) => Promise<void>;
-
-  /**
-   * Withdraw rewards from all validators. The promise settles when the rewards are withdrawn.
-   * @returns The total amounts of rewards withdrawn
-   */
-  withdrawRewards: () => Promise<ChainAmount[]>;
-
-  /**
-   * Withdraw rewards from a specific validator. The promise settles when the rewards are withdrawn.
-   * @param validator - the validator to withdraw rewards from
-   * @returns
-   */
-  withdrawReward: (validator: CosmosValidatorAddress) => Promise<ChainAmount[]>;
-}
-
-/**
  * An object that supports high-level operations for an account on a remote chain.
  */
 export interface BaseOrchestrationAccount
   extends StakingAccountQueries,
     StakingAccountActions {
   /** @returns the underlying low-level operation object. */
-  getChainAcccount: () => Promise<ChainAccount>;
+  getChainAcccount: () => Promise<IcaAccount>;
 
   /**
    * @returns the address of the account on the remote chain
